Eto.Gtk: 适应Mac

+++ TsinghuaNet.Eto/TsinghuaNet.Eto/Converters/FluxLimitConverter.cs
@@ -9,12 +9,12 @@ namespace TsinghuaNet.Eto.Converters
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
             ByteSize bytes = (ByteSize)value;
-            return (int)bytes.GigaBytes;
+            return bytes.GigaBytes;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return ByteSize.FromGigaBytes((int)value);
+            return ByteSize.FromGigaBytes((double)value);
         }
     }
 }
